Implement mergesort from java and fix memory leaks
This should have been 3 commit in all honesty...
First off the code was brought in from java code and was translated. Issues with that were fixed ie. fixing array decleration.
Another big change was switching around the order of the functions so that the code would be able to compile.
And lastly destination was patched